% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_concept_helpers.R
\name{broaden_concept}
\alias{broaden_concept}
\title{Function to add a variable with broader concepts to a dataframe}
\usage{
broaden_concept(data, uri, new_var)
}
\arguments{
\item{data}{a dataframe containing a variable with concept uri's.}

\item{uri}{the dataframe variable containing the uri of the concept for
which broader concepts should be added.}

\item{new_var}{name of the new variable to be added to the dataframe}
}
\value{
A dataframe with an added variable containing nested dataframes with
 variables 'uri' and 'label'. The nested datasets include the original
 concepts.
}
\description{
`broaden_concept()` Given a dataset wich containing a column containing
  Flanders Heritage concept URI's, this function will add a column containing
  a nested data frame with the URI and label of this concept and all broader
  concepts. The primary goal is to allow for easy filtering on specific
  concepts by URI or Label while taking all narrower labels into account as
  well. To do this, the dataset can either be unnested using tidyr::unnest()
  or by using the filter_concept() function of this package.
}
\examples{
broaden_concept(d, uri, broader_concepts)
}
