% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_concept_helpers.R
\name{get_broader_concepts}
\alias{get_broader_concepts}
\title{Function to get all broader concepts of a specific concept uri}
\usage{
get_broader_concepts(uri)
}
\arguments{
\item{uri}{a character string containg the uri of a concept}
}
\value{
A dataframe with two variables (uri, label), containing the
 uri's and labels of the input concept and all broader concepts.
}
\description{
`get_broader_concepts()` get a dataset containing the broader concepts of
a specific concept by uri.
}
\examples{
get_broader_concepts('https://id.erfgoed.net/thesauri/dateringen/1196')
}
