% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_concept_helpers.R
\name{get_concept}
\alias{get_concept}
\title{Function to get the direct broader concept of a specific concept uri}
\usage{
get_concept(uri)
}
\arguments{
\item{uri}{a character string containg the uri of a concept}
}
\value{
An object representing the parsed json body of the concept.
}
\description{
`get_concept()` get the parsed json body returned from resolving the
 concept uri.
}
\examples{
get_concept('https://id.erfgoed.net/thesauri/dateringen/1196')
}
