% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oe_eindverslagen.R
\name{get_oe_eindverslagen}
\alias{get_oe_eindverslagen}
\title{Get Flanders Heritage layers containing archaeological reports (eindverslagen)}
\usage{
get_oe_eindverslagen()
}
\value{
An object of class sf (see \link[sf]{st_read}). The object has a
comment containing the retrieval_time and retrieval_url.
}
\description{
`get_oe_eindverslagen()` get an OGC Simple Features collection
containing the final reports of archeological projects submitted
to Flanders Heritage.
}
\examples{
get_oe_eindverslagen()
}
