% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oe_layer.R
\name{get_oe_layer}
\alias{get_oe_layer}
\title{Get Flanders Heritage layer containing}
\usage{
get_oe_layer(download_filename, layer)
}
\arguments{
\item{download_filename}{Filename of the zip file containing the layer to
be extracted, as mentioned on the download page \url{https://geo.onroerenderfgoed.be/downloads/}.
e.g. 'archeologie.zip}

\item{layer}{Name of the layer (shapefile) to be extracted. e.g.
eindverslagen.shp}
}
\value{
An object of class sf (see \link[sf]{st_read}). The object has a
  comment containing the retrieval_time and retrieval_url.
}
\description{
`get_oe_layer()` get an OGC Simple Features collection
containing the specified layer. The package also provides wrapper functions
for different layers. It is recommended to use the wrapper functions, as this
function is dependent on the implementation of the download service provided
by Flanders heritage. i.e. the download file name has to be passed in by
calling code.
}
\examples{
get_oe_layer('archeologie.zip', 'eindverslagen')
}
