% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oe_notas.R
\name{get_oe_notas}
\alias{get_oe_notas}
\title{Get Flanders Heritage dataset of archaeology memorandums and memorandums (archeologienotas en notas)}
\usage{
get_oe_notas()
}
\value{
An object of class sf (see \link[sf]{st_read}). The object has a
comment containing the retrieval_time and retrieval_url.
}
\description{
`get_oe_notas()` get an OGC Simple Features collection
containing archaeology memorandums (archeologienotas) and memorandums (notas)
resulting from preliminary archaeological research submitted
to Flanders Heritage.
}
\examples{
get_oe_notas()
}
