% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vrbg_collection_ids.R
\name{get_vrbg_collection_ids}
\alias{get_vrbg_collection_ids}
\title{Get collection ID's of administrative units in Flanders}
\usage{
get_vrbg_collection_ids(version = "VRBG")
}
\arguments{
\item{version}{a string representing the version of the reference file.
Default is 'VRBG', which is always the most recent version. specific
versions are also possible (e.g. 'VRBG2003').}
}
\value{
character vector containing collection id's.
}
\description{
`get_vrbg_collection_ids()` get an character vector of the ID's of
collections present in a  specific version of the Provisional Reference File
of Municipal Borders (Voorlopig Referentiebestand Gemeentegrenzen) as
published through the OGC Features API of the Flemish Government (see:
https://metadata.vlaanderen.be/srv/dut/catalog.search#/metadata/3c1e59e6-8a28-46be-b31a-eb838376cc65).
}
\examples{
get_vrbg_collection_ids()
get_vrbg_collection_ids('VRBG2003')
get_vrbg_collection_ids(version = 'VRBG2003')
}
