% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vrbg_sf.R
\name{get_vrbg_sf}
\alias{get_vrbg_sf}
\title{Get reference datasets for administrative borders in Flanders}
\usage{
get_vrbg_sf(collection = "Refgem", version = "VRBG", crs = "EPSG:31370")
}
\arguments{
\item{collection}{a string representing the
collection (i.e. type of administrative unit) that should be returned.
Possible values for the default version of the reference data set are:
\itemize{
  \item{Refgem}{municipalities, default}
  \item{'Refgew'}{Flemish region}
  \item{Refprv}{}{provinces}
  \item{'Refarr'}{arrondissements}
}
The collection names are case-sensitive.}

\item{version}{a string representing the version of the reference file.
Default is 'VRBG', which is always the most recent version. specific
versions are also possible (e.g. 'VRBG2003'). Note that not all versions
contain the same collections.}

\item{crs}{a string representing the EPSG code for the coordinate reference
system. Default value is 'EPSG:31370' for Belgian Lambert 1972. Other likely
values are 'EPSG:4326' (WGS84), 'EPSG:3812' (Belgian Lambert 2008)
and 'EPSG:4258' (ETRS89). For a complete list of supported reference systems,
inspect the collections endpoint.}
}
\value{
An object of class sf (see \link[sf]{st_read}). The object has a
  comment containing the retrieval_time and retrieval_url.
}
\description{
`get_vrbg_sf()` Get an OGC Simple Features collection that is
  part of the Provisional Reference File of Municipal Borders (Voorlopig
  Referentiebestand Gemeentegrenzen) as published through the OGC Features API
  of the Flemish Government (see:
  \url{https://metadata.vlaanderen.be/srv/dut/catalog.search#/metadata/3c1e59e6-8a28-46be-b31a-eb838376cc65}).
}
\examples{
get_vrbg_sf()
get_vrbg_sf('Refarr')
get_vrgb_sf(collection = 'Refarr', crs = 'EPSG:3812')
}
