% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vrbg_supported_crs.R
\name{get_vrbg_supported_crs}
\alias{get_vrbg_supported_crs}
\title{Get supported CRS for administrative borders in Flanders}
\usage{
get_vrbg_supported_crs(version = "VRBG")
}
\arguments{
\item{version}{a string representing the version of the reference file.
Default is 'VRBG', which is always the most recent version. specific
versions are also possible (e.g. 'VRBG2003').}
}
\value{
character vector containing EPSG codes in the format 'EPSG:<number>'.
}
\description{
`get_vrbg_supported_crs()` get an character vector of the EPSG codes of
coordinate reference systems supported byspecific version of the
Provisional Reference File of Municipal Borders (Voorlopig Referentiebestand
Gemeentegrenzen) as published through the OGC Features API of the
Flemish Government (see:
\href{https://metadata.vlaanderen.be/srv/dut/catalog.search#/metadata/3c1e59e6-8a28-46be-b31a-eb838376cc65}{metadata catalog entry})
}
\examples{
get_vrbg_supported_crs()
get_vrbg_supported_crs('VRBG2003')
get_vrbg_supported_crs(version = 'VRBG2003')
}
