% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_from_geojson.R
\name{resp_sf_from_geojson}
\alias{resp_sf_from_geojson}
\title{Extract sf object from httr2 response object containing geojson}
\usage{
resp_sf_from_geojson(resp, annotated = TRUE)
}
\arguments{
\item{resp}{a httr2 response object}

\item{annotated}{a boolean value (default TRUE). When TRUE, the returned
object contains metdadata (retrieval_url and retrieval_time) in the object's
comment.}
}
\value{
An object of class sf (see \link[sf]{st_read}). If the annotated
parameter is TRUE (default), the returned object will have a
comment containing the retrieval_time and retrieval_url.
}
\description{
`resp_sf_from_geojson` This function extracts a Simple Features (sf) object
from an httr2 response object when the response body contains a feature set
in geojson format.
}
\examples{
\dontrun{
resp_from_geojson(resp)
resp_from_geojson(resp, annotated=TRUE)
}
}
