% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_concept_helpers.R
\name{broaden_concept}
\alias{broaden_concept}
\title{Function to add a variable with broader concepts to a dataframe}
\usage{
broaden_concept(data, uri, new_var)
}
\arguments{
\item{data}{a dataframe containing a variable with concept uri's.}

\item{uri}{the dataframe variable containing the uri of the concept for
which broader concepts should be added.}

\item{new_var}{name of the new variable to be added to the dataframe}
}
\value{
A dataframe with an added variable containing nested dataframes with
 variables 'uri' and 'label'. The nested datasets include the original
 concepts.
}
\description{
`broaden_concept()` Filter a dataset on a column with concepts as nested
  dataframes. The nested dataframes must contain a 'label' variable and a
  'uri' variable. This is the format returned by the broaden_concepts
  function.
}
\examples{
broaden_concept(d, uri, broader concepts)
}
