% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_concept_helpers.R
\name{filter_concept}
\alias{filter_concept}
\title{Function to filter a dataset based with nested concepts on concept label}
\usage{
filter_concept(data, var, concept)
}
\arguments{
\item{data}{a dataframe containing a variable with concept uri's.}

\item{concept}{a string representing the uri or label of the concept that
should be used to filter the dataset. The filter
function will return all rows with an exact match for this label or URI in
the nested dataset.}

\item{uri}{the dataframe variable containing the nested data frames}
}
\value{
A filtered data frame
}
\description{
`filter_concept()` Filter a dataset on a column with concepts as nested
  dataframes. The nested dataframes must contain a 'label' variable and a
  'uri' variable. This is the format returned by the broaden_concepts
  function.
}
\examples{
filter(d, concepts, 'ijzertijd')
filter(d, concepts, 'https://id.erfgoed.net/thesauri/dateringen/1293')
}
